+++ client/package.json
@@ -1,6 +1,6 @@
 {
-  "name": "tunji-web-deux",
-  "version": "0.1.0",
+  "name": "client",
+  "version": "1.0.0",
   "private": true,
   "proxy": "https://localhost:8080",
   "dependencies": {
@@ -43,6 +43,7 @@
     "remark-gfm": "^1.0.0",
     "reselect": "^4.0.0",
     "typescript": "^4.1.2",
+    "common": "^1.0.0",
     "web-vitals": "^1.0.1"
   },
   "scripts": {
