+++ common/package.json
@@ -1,0 +1,7 @@
+{
+  "name": "common",
+  "version": "1.0.0",
+
+  "dependencies": {
+  }
+}
