+++ server/package.json
@@ -1,0 +1,66 @@
+{
+  "name": "server",
+  "version": "1.0.0",
+  "private": true,
+  "scripts": {
+    "build-client": "cd client; yarn install; npm run build; cd ..;",
+    "copy-ssl": "cp domain.crt build/domain.crt; cp domain.key build/domain.key",
+    "build-ts": "tsc --noEmit false && npm run copy-ssl",
+    "build": "npm run build-ts && npm run lint",
+    "build-all": "npm run build && npm run build-client",
+    "debug": "npm run build && npm run watch-debug",
+    "lint": "tsc --noEmit && eslint \"**/*.{js,ts}\" --quiet --fix",
+    "serve-debug": "nodemon --inspect dist/server.js",
+    "serve": "node build/server.js",
+    "start": "npm run serve"
+  },
+  "engines": {
+    "node": "12.x"
+  },
+  "dependencies": {
+    "@types/bluebird": "^3.5.33",
+    "@types/compression": "^1.7.0",
+    "@types/connect-mongo": "^3.1.3",
+    "@types/cookie-parser": "^1.4.2",
+    "@types/cors": "^2.8.10",
+    "@types/express": "^4.17.11",
+    "@types/express-session": "^1.17.3",
+    "@types/helmet": "^4.0.0",
+    "@types/mongoose": "^5.10.5",
+    "@types/morgan": "^1.9.2",
+    "@types/open-graph-scraper": "^4.7.0",
+    "@types/passport": "^1.0.6",
+    "@types/passport-local": "^1.0.33",
+    "@types/serve-favicon": "^2.5.2",
+    "bluebird": "^3.5.4",
+    "body-parser": "^1.18.3",
+    "compression": "^1.7.4",
+    "connect-mongo": "^4.4.1",
+    "cookie-parser": "^1.4.4",
+    "cors": "^2.8.5",
+    "debug": "^4.3.1",
+    "ejs": "*",
+    "express": "^4.16.4",
+    "express-session": "^1.17.1",
+    "helmet": "^4.5.0",
+    "mongoose": "^5.4.20",
+    "morgan": "^1.9.1",
+    "nodemailer": "6.5.0",
+    "open-graph-scraper": "^4.8.2",
+    "passport": "^0.4.1",
+    "passport-local": "^1.0.0",
+    "pm2": "^4.5.6",
+    "rate-limiter-flexible": "^2.2.2",
+    "serve-favicon": "^2.5.0",
+    "client": "^1.0.0",
+    "common": "^1.0.0"
+  },
+  "devDependencies": {
+    "@types/eslint": "^7.2.6",
+    "@types/node": "^14.14.25",
+    "@typescript-eslint/eslint-plugin": "^4.14.2",
+    "@typescript-eslint/parser": "^4.14.2",
+    "eslint": "^7.19.0",
+    "typescript": "^4.1.3"
+  }
+}
